Refactor input checks to separate file
